# src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Express.Multer.File": {
				"description": "Object containing file metadata and access information.",
				"properties": {
					"fieldname": {
						"type": "string",
						"description": "Name of the form field associated with this file."
					},
					"originalname": {
						"type": "string",
						"description": "Name of the file on the uploader's computer."
					},
					"encoding": {
						"type": "string",
						"description": "Value of the `Content-Transfer-Encoding` header for this file.",
						"deprecated": true
					},
					"mimetype": {
						"type": "string",
						"description": "Value of the `Content-Type` header for this file."
					},
					"size": {
						"type": "number",
						"format": "double",
						"description": "Size of the file in bytes."
					},
					"stream": {
						"type": "string",
						"format": "byte",
						"description": "A readable stream of this file. Only available to the `_handleFile`\ncallback for custom `StorageEngine`s."
					},
					"destination": {
						"type": "string",
						"description": "`DiskStorage` only: Directory to which this file has been uploaded."
					},
					"filename": {
						"type": "string",
						"description": "`DiskStorage` only: Name of this file within `destination`."
					},
					"path": {
						"type": "string",
						"description": "`DiskStorage` only: Full path to the uploaded file."
					},
					"buffer": {
						"type": "string",
						"format": "byte",
						"description": "`MemoryStorage` only: A Buffer containing the entire file."
					}
				},
				"required": [
					"fieldname",
					"originalname",
					"encoding",
					"mimetype",
					"size",
					"stream",
					"destination",
					"filename",
					"path",
					"buffer"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateIdeaResponse": {
				"properties": {
					"idea": {
						"properties": {
							"updated_at": {
								"type": "string"
							},
							"created_at": {
								"type": "string"
							},
							"links": {
								"items": {
									"properties": {
										"url": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"url",
										"id"
									],
									"type": "object"
								},
								"type": "array"
							},
							"images": {
								"items": {
									"properties": {
										"height": {
											"type": "number",
											"format": "double"
										},
										"width": {
											"type": "number",
											"format": "double"
										},
										"alt": {
											"type": "string"
										},
										"src": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"height",
										"width",
										"alt",
										"src",
										"id"
									],
									"type": "object"
								},
								"type": "array"
							},
							"tags": {
								"items": {
									"properties": {
										"name": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"name",
										"id"
									],
									"type": "object"
								},
								"type": "array"
							},
							"authorId": {
								"type": "string"
							},
							"description": {
								"type": "string",
								"nullable": true
							},
							"title": {
								"type": "string",
								"nullable": true
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"updated_at",
							"created_at",
							"authorId",
							"description",
							"title",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"idea"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateIdeaInput": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/Express.Multer.File"
						},
						"type": "array"
					},
					"links": {
						"items": {
							"properties": {
								"url": {
									"type": "string"
								}
							},
							"required": [
								"url"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GetIdeaByIdResponse": {
				"properties": {
					"idea": {
						"properties": {
							"_count": {
								"properties": {
									"comments": {
										"type": "number",
										"format": "double"
									},
									"views": {
										"type": "number",
										"format": "double"
									},
									"likes": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"comments",
									"views",
									"likes"
								],
								"type": "object"
							},
							"updated_at": {
								"type": "string"
							},
							"created_at": {
								"type": "string"
							},
							"author": {
								"properties": {
									"icon": {
										"type": "string",
										"nullable": true
									},
									"displayName": {
										"type": "string",
										"nullable": true
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							},
							"authorId": {
								"type": "string"
							},
							"description": {
								"type": "string",
								"nullable": true
							},
							"title": {
								"type": "string",
								"nullable": true
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"updated_at",
							"created_at",
							"author",
							"authorId",
							"description",
							"title",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"idea"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateIdeaResponse": {
				"properties": {
					"idea": {
						"properties": {
							"updated_at": {
								"type": "string"
							},
							"created_at": {
								"type": "string"
							},
							"links": {
								"items": {
									"properties": {
										"url": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"url",
										"id"
									],
									"type": "object"
								},
								"type": "array"
							},
							"images": {
								"items": {
									"properties": {
										"height": {
											"type": "number",
											"format": "double"
										},
										"width": {
											"type": "number",
											"format": "double"
										},
										"alt": {
											"type": "string"
										},
										"src": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"height",
										"width",
										"alt",
										"src",
										"id"
									],
									"type": "object"
								},
								"type": "array"
							},
							"tags": {
								"items": {
									"properties": {
										"name": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"name",
										"id"
									],
									"type": "object"
								},
								"type": "array"
							},
							"authorId": {
								"type": "string"
							},
							"description": {
								"type": "string",
								"nullable": true
							},
							"title": {
								"type": "string",
								"nullable": true
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"updated_at",
							"created_at",
							"authorId",
							"description",
							"title",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"idea"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateIdeaInput": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/Express.Multer.File"
						},
						"type": "array"
					},
					"links": {
						"items": {
							"properties": {
								"url": {
									"type": "string"
								}
							},
							"required": [
								"url"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FollowingResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"followingUser": {
						"properties": {
							"username": {
								"type": "string"
							},
							"icon": {
								"type": "string",
								"nullable": true
							},
							"displayName": {
								"type": "string",
								"nullable": true
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					},
					"created_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"followingUser",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetFollowingResponse": {
				"properties": {
					"following": {
						"items": {
							"$ref": "#/components/schemas/FollowingResponse"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"hasPrev": {
								"type": "boolean"
							},
							"hasNext": {
								"type": "boolean"
							},
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"hasPrev",
							"hasNext",
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"following",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FollowerResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"followerUser": {
						"properties": {
							"username": {
								"type": "string"
							},
							"icon": {
								"type": "string",
								"nullable": true
							},
							"displayName": {
								"type": "string",
								"nullable": true
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					},
					"created_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"followerUser",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetFollowersResponse": {
				"properties": {
					"followers": {
						"items": {
							"$ref": "#/components/schemas/FollowerResponse"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"hasPrev": {
								"type": "boolean"
							},
							"hasNext": {
								"type": "boolean"
							},
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"hasPrev",
							"hasNext",
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"followers",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
...
		"/v1/ideas/{id}": {
			"patch": {
				"operationId": "UpdateIdea",
				"responses": {
					"200": {
						"description": "Idea updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UpdateIdeaResponse"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Update an existing idea\n\nThis endpoint allows authenticated users to update their own ideas.\nOnly the author of the idea can update it.",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the idea to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The request body containing the updated idea details.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateIdeaInput",
								"description": "The request body containing the updated idea details."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetIdeaById",
				"responses": {
					"200": {
						"description": "Idea retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/GetIdeaByIdResponse"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Get a specific idea by ID\n\nThis endpoint allows authenticated users to retrieve a specific idea by its ID.\nThe response includes the idea details, author information, and engagement metrics.",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the idea to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteIdea",
				"responses": {
					"200": {
						"description": "Idea deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Delete an existing idea\n\nThis endpoint allows authenticated users to delete their own ideas.\nOnly the author of the idea can delete it. This action is irreversible.",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the idea to delete.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/ideas": {
			"get": {
				"operationId": "GetIdeas",
				"responses": {
					"200": {
						"description": "Ideas retrieved successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get list of ideas with filtering and pagination\n\nThis endpoint allows authenticated users to retrieve a paginated list of ideas\nwith various filtering options. Users can search by title/description, filter\nby author, tags, and sort by different criteria.\n\nIdeas Retrieval Process:\n- Validates query parameters (pagination, filters, sorting)\n- Applies search filters if provided\n- Retrieves ideas from database with pagination\n- Includes author information and engagement metrics\n- Returns paginated results with metadata",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "authorId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"created_at",
								"updated_at",
								"title"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "CreateIdea",
				"responses": {
					"201": {
						"description": "Idea created successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CreateIdeaResponse"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Create a new idea\n\nThis endpoint allows authenticated users to create new ideas. The idea will be\nassociated with the authenticated user's profile and will be visible to other\nusers based on the platform's visibility rules.\n\nIdea Creation Process:\n- Validates input data (title, description, summary, tags)\n- Associates idea with authenticated user's profile\n- Creates idea record in database\n- Returns created idea with generated ID and timestamps",
				"tags": [
					"Ideas"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- Object containing idea data (title, description, summary, tags)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateIdeaInput",
								"description": "- Object containing idea data (title, description, summary, tags)"
							}
						}
					}
				}
			}
		},
...
		"/v1/follow/following": {
			"get": {
				"operationId": "GetFollowing",
				"responses": {
					"200": {
						"description": "Following retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/GetFollowingResponse"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Get who a user is following\n\nThis endpoint allows authenticated users to retrieve a paginated list of users\nthat a specific user profile is following.",
				"tags": [
					"Follows"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- The user profile ID to get following for",
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "- Page number (0-based)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Number of items per page",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/follow/followers": {
			"get": {
				"operationId": "GetFollowers",
				"responses": {
					"200": {
						"description": "Followers retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/GetFollowersResponse"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Get followers of a user\n\nThis endpoint allows authenticated users to retrieve a paginated list of followers\nfor a specific user profile. Users can see who is following them.",
				"tags": [
					"Follows"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- The user profile ID to get followers for",
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "- Page number (0-based)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Number of items per page",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}